+++ lib/ui/canvas/CanvasCourseHomePanel.dart
@@ -17,27 +17,20 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:illinois/model/Canvas.dart';
-import 'package:illinois/service/Config.dart';
-import 'package:illinois/ui/canvas/CanvasAccountNotificationsPanel.dart';
-import 'package:illinois/ui/canvas/CanvasCourseAssignmentsPanel.dart';
-import 'package:illinois/ui/canvas/CanvasCourseModulesPanel.dart';
-import 'package:illinois/ui/canvas/CanvasSyllabusHtmlPanel.dart';
-import 'package:illinois/ui/groups/GroupsHomePanel.dart';
 import 'package:illinois/service/Analytics.dart';
 import 'package:illinois/service/Canvas.dart';
-import 'package:illinois/ui/canvas/CanvasCourseAnnouncementsPanel.dart';
-import 'package:illinois/ui/canvas/CanvasCourseCalendarPanel.dart';
-import 'package:illinois/ui/canvas/CanvasCourseCollaborationsPanel.dart';
-import 'package:illinois/ui/canvas/CanvasFileSystemEntitiesListPanel.dart';
+import 'package:illinois/service/Config.dart';
+import 'package:illinois/ui/canvas/CanvasCourseAssignmentsPanel.dart';
 import 'package:illinois/ui/canvas/CanvasWidgets.dart';
-import 'package:illinois/ui/inbox/InboxHomePanel.dart';
 import 'package:illinois/ui/widgets/HeaderBar.dart';
 import 'package:illinois/ui/widgets/RibbonButton.dart';
 import 'package:illinois/ui/widgets/TabBar.dart' as uiuc;
+import 'package:rokwire_plugin/rokwire_plugin.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 import 'package:rokwire_plugin/service/styles.dart';
 import 'package:rokwire_plugin/utils/utils.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:sprintf/sprintf.dart';
+import 'package:url_launcher/url_launcher.dart' as url_launcher;
 
 class CanvasCourseHomePanel extends StatefulWidget {
   final int? courseId;
@@ -101,78 +94,17 @@ class _CanvasCourseHomePanelState extends State<CanvasCourseHomePanel> {
     return Column(children: [
       _buildDelimiter(),
       RibbonButton(
-          label: Localization().getStringEx('panel.home_canvas_course.button.syllabus.title', 'Syllabus'),
-          hint: Localization().getStringEx('panel.home_canvas_course.button.syllabus.hint', ''),
-          leftIconAsset: 'images/icon-canvas-implemented-working.png',
-          onTap: _onTapSyllabus),
-      _buildDelimiter(),
-      RibbonButton(
-          label: Localization().getStringEx('panel.home_canvas_course.button.announcements.title', 'Announcements'),
-          hint: Localization().getStringEx('panel.home_canvas_course.button.announcements.hint', ''),
-          leftIconAsset: 'images/icon-canvas-implemented-working.png',
-          onTap: _onTapAnnouncements),
-      _buildDelimiter(),
-      RibbonButton(
-          label: Localization().getStringEx('panel.home_canvas_course.button.files.title', 'Files'),
-          hint: Localization().getStringEx('panel.home_canvas_course.button.files.hint', ''),
-          leftIconAsset: 'images/icon-canvas-implemented-working.png',
-          onTap: _onTapFiles),
-      _buildDelimiter(),
-      RibbonButton(
-          label: Localization().getStringEx('panel.home_canvas_course.button.collaborations.title', 'Collaborations'),
-          hint: Localization().getStringEx('panel.home_canvas_course.button.collaborations.hint', ''),
-          leftIconAsset: 'images/icon-canvas-implemented-working.png',
-          onTap: _onTapCollaborations),
-      _buildDelimiter(),
-      RibbonButton(
-          label: Localization().getStringEx('panel.home_canvas_course.button.calendar.title', 'Calendar'),
-          hint: Localization().getStringEx('panel.home_canvas_course.button.calendar.hint', ''),
-          leftIconAsset: 'images/icon-canvas-implemented-working.png',
-          onTap: _onTapCalendar),
-      _buildDelimiter(),
-      RibbonButton(
-          label: Localization().getStringEx('panel.home_canvas_course.button.notifications.title', 'Notifications'),
-          hint: Localization().getStringEx('panel.home_canvas_course.button.notifications.hint', ''),
-          leftIconAsset: 'images/icon-canvas-implemented-working.png',
-          onTap: _onTapNotifications),
-      _buildDelimiter(),
-      RibbonButton(
-          label: Localization().getStringEx('panel.home_canvas_course.button.notitification.history.title', 'Notification history'),
-          hint: Localization().getStringEx('panel.home_canvas_course.button.notitification.history.hint', ''),
-          leftIconAsset: 'images/icon-canvas-implemented-working.png',
-          onTap: _onTapNotificationHistory),
-      _buildDelimiter(),
-      RibbonButton(
-          label: Localization().getStringEx('panel.home_canvas_course.button.modules.title', 'Modules'),
-          hint: Localization().getStringEx('panel.home_canvas_course.button.modules.hint', ''),
-          leftIconAsset: 'images/icon-canvas-implemented-working.png',
-          onTap: _onTapModules),
-      _buildDelimiter(),
-      RibbonButton(
           label: Localization().getStringEx('panel.home_canvas_course.button.assignments.title', 'Assignments'),
           hint: Localization().getStringEx('panel.home_canvas_course.button.assignments.hint', ''),
           leftIconAsset: 'images/icon-canvas-implemented-working.png',
           onTap: _onTapAssignments),
       _buildDelimiter(),
       RibbonButton(
-          label: Localization().getStringEx('panel.home_canvas_course.button.zoom.title', 'Zoom Meeting'),
-          hint: Localization().getStringEx('panel.home_canvas_course.button.zoom.hint', ''),
+          label: Localization().getStringEx('panel.home_canvas_course.button.launch.title', 'Launch Canvas'),
+          hint: Localization().getStringEx('panel.home_canvas_course.button.launch.hint', ''),
           leftIconAsset: 'images/icon-canvas-implemented-working.png',
-          onTap: _onTapZoomMeeting),
-      _buildDelimiter(),
-      RibbonButton(
-          label: Localization().getStringEx('panel.home_canvas_course.button.group.title', 'Group'),
-          hint: Localization().getStringEx('panel.home_canvas_course.button.group.hint', ''),
-          leftIconAsset: 'images/icon-canvas-implemented-working.png',
-          onTap: _onTapGroup),
-      _buildDelimiter(),
-      /* Hide feedback for now
-      RibbonButton(
-          label: Localization().getStringEx('panel.home_canvas_course.button.feedback.title', 'Feedback'),
-          hint: Localization().getStringEx('panel.home_canvas_course.button.feedback.hint', ''),
-          leftIconAsset: 'images/icon-canvas-implemented-working.png',
-          onTap: _onTapFeedback),
-      _buildDelimiter()*/
+          onTap: _onTapLaunch),
+      _buildDelimiter()
     ]);
   }
 
@@ -180,69 +112,23 @@ class _CanvasCourseHomePanelState extends State<CanvasCourseHomePanel> {
     return Container(height: 1, color: Styles().colors!.surfaceAccent);
   }
 
-  void _onTapSyllabus() {
-    Analytics().logSelect(target: 'Canvas Course -> Syllabus');
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasSyllabusHtmlPanel(courseId: widget.courseId!)));
-  }
-
-  void _onTapAnnouncements() {
-    Analytics().logSelect(target: 'Canvas Course -> Announcements');
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasCourseAnnouncementsPanel(courseId: widget.courseId!)));
-  }
-
-  void _onTapFiles() {
-    Analytics().logSelect(target: 'Canvas Course -> Files');
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasFileSystemEntitiesListPanel(courseId: widget.courseId)));
-  }
-
-  void _onTapCollaborations() {
-    Analytics().logSelect(target: 'Canvas Course -> Collaborations');
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasCourseCollaborationsPanel(courseId: widget.courseId!)));
-  }
-
-  void _onTapCalendar() {
-    Analytics().logSelect(target: 'Canvas Course -> Calendar');
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasCourseCalendarPanel(courseId: widget.courseId!)));
-  }
-
-  void _onTapNotifications() {
-    Analytics().logSelect(target: 'Canvas Course -> Notifications');
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasAccountNotificationsPanel()));
-  }
-
-  void _onTapNotificationHistory() {
-    Analytics().logSelect(target: 'Canvas Course -> Notification history');
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => InboxHomePanel()));
-  }
-
-  void _onTapModules() {
-    Analytics().logSelect(target: 'Canvas Course -> Modules');
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasCourseModulesPanel(courseId: widget.courseId!)));
+  void _onTapLaunch() async {
+    Analytics().logSelect(target: 'Canvas Course -> Launch Canvas');
+    String? courseDeepLinkFormat = Config().canvasCourseDeepLinkFormat;
+    String? courseDeepLink = StringUtils.isNotEmpty(courseDeepLinkFormat) ? sprintf(courseDeepLinkFormat!, [_course!.id]) : null;
+    bool? appLaunched = await RokwirePlugin.launchApp({"deep_link": courseDeepLink});
+    if (appLaunched != true) {
+      String? canvasStoreUrl = Config().canvasStoreUrl;
+      if ((canvasStoreUrl != null) && await url_launcher.canLaunch(canvasStoreUrl)) {
+        await url_launcher.launch(canvasStoreUrl, forceSafariVC: false);
+      }
+    }
   }
 
   void _onTapAssignments() {
     Analytics().logSelect(target: 'Canvas Course -> Assignments');
     Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasCourseAssignmentsPanel(courseId: widget.courseId!)));
   }
-
-  void _onTapGroup() {
-    Analytics().logSelect(target: 'Canvas Course -> Group');
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => GroupsHomePanel()));
-  }
-
-  void _onTapZoomMeeting() {
-    Analytics().logSelect(target: 'Canvas Course -> Zoom Meeting');
-    String? zoomUrl = Config().canvasZoomMeetingUrl;
-    if (StringUtils.isNotEmpty(zoomUrl)) {
-      launch(zoomUrl!);
-    }
-  }
-
-  /* Hide Feedback for now
-  void _onTapFeedback() {
-    Analytics().logSelect(target: 'Canvas Course -> Feedback');
-    Navigator.push(context, CupertinoPageRoute(builder: (context) => CanvasFeedbackPanel()));
-  }*/
 
   void _loadCourse() {
     _increaseProgress();