+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Launch "Canvas Student" app - deep link to specific assignment [#1661](https://github.com/rokwire/illinois-app/issues/1661).
 
 ## [4.0.15] - 2022-05-23
 ### Changed